Factor out builder
